Move most source files to src subdirectory
